pyproject.toml
[project]
name = "auto-rocket-gui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.2",
    "ddddocr>=1.5.6",
    "keyring>=25.6.0",
    "pillow>=12.0.0",
    "python-socks>=2.7.2",
    "sqlalchemy>=2.0.44",
    "websockets[socks]>=15.0.1",
]
